instantiates, runs, and plots a cmap of an ECA
Currently this runs rule 110 for 11250 generations taking in an initial state of 10000 randomly generated binary digits. It then plots a blue-green colormap of these generations. The boundary condition is periodic, so the last cell is counted to the left of the first and the first is counted to the right of the last for the purpose of calculating the cells directly beneath the first and last. One could visualize colormap as being printed on a sheet of paper and folded into a cylinder connecting the ends of each generation to get a feel for what this boundary effect is doing. I plan to add a user interface which allows for easy and intuitive modification of all desired parameters, but currently this can be done within the code itself via changing the inputs into the various methods and functions.